Move text formatting function to separate utility file following single responsibility principle
