+++ components/Chat/TextMessagePart.tsx
@@ -5,24 +5,11 @@ import remarkGfm from 'remark-gfm';
 import rehypeRaw from 'rehype-raw';
 import { TextPart } from "@/types/reasoning";
 import styles from "./markdown.module.css";
+import { convertIndentToBullet } from "@/lib/utils/textFormatting";
 
 interface TextMessagePartProps {
   part: TextPart;
 }
-
-// Converts indented lines to proper markdown bullet lists
-const convertIndentToBullet = (text: string): string => {
-  if (!text) return '';
-  
-  const lines = text.split('\n');
-  
-  return lines.map(line => {
-    if (/^\s{2,}[A-Za-z]/.test(line) && !line.trim().startsWith('*') && !line.trim().startsWith('-')) {
-      return line.replace(/^\s+/, '* ');
-    }
-    return line;
-  }).join('\n');
-};
 
 export function TextMessagePart({ part }: TextMessagePartProps) {
   const processedText = convertIndentToBullet(part.text);
